pyproject.toml
[project]
name = "a-new-era-of-agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "finnhub-python>=2.4.23",
    "google-adk>=0.1.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "termcolor>=3.0.1",
]
